Switched FEproblem().elements to a set and added get_nodes().
I figured that, if nodes are based off object identity, then elements should be
too.  Otherwise, duplicate elements would appear in the output if inadvertently
added to the list multiple times.  Modified test_formats to use this.

In the process, created a simple FEproblem method to get the set of nodes used
by elements in the problem.  Modified feb to use this.

Also added real tests with real XML parsing to test_formats.