Created filetype-detecting read and write methods for FEproblem.
They currently just check the file extension, then call "read_{ext}" or
"write_{ext}", but it does the trick.  In order to work with this, all readers
and writers need the same signature.  Therefore, the .inp reader was changed to
accept a filename instead of a file object.  This involved an extra layer of
indentation, so many lines have changed, even though little code did.