Added support for materials in write_feb.
It's pretty massive, involving a bevy of hidden methods that get added to
material classes in order to help it along.  Consequently, the same thing was
done for geometry classes, eliminating the element_write_map used before.

Implementing everything also involved a small fix to FEproblem.get_materials to
ignore Nones, and a fix to test_materials.py to reflect the replacement of the
parameters dict with a list.  And while we're at it, a single comment fix to
geometry.py.