Finalized approach for switching (aka steps in FEBio).
This approach is rather different from FEBio's.  Rather than having specific
steps, objects can instead be wrapped in a derivative of the Switch class.  It
gets used as a mixin, allowing it to present itself as other types (eg. one
that acts as a Constraint), and contains a series of that type, each mapped to
a specific time.

As currently implemented, it's only useful for Constraints, allowing different
constraints to activate or deactivate at different times.  It will also be used
for Contact, and who knows what else.