Loadcurve output in feb files, plus supporting FEproblem methods.
This is the start of constraint output in feb format.  The constraints
themselves must still be output, but all their loadcurves are supported.  But
no tests for this yet.

This also necessitated creating three new methods for the FEproblem class:
get_constrainables, get_constraints, and get_loadcurves.  Might also need to
create get_switches.  But all these methods are yet untested.