All objects now inherit from Base.  AxisOrientation objects are therefore required.
Having the common Base now allows the entire FEproblem structure to be
traversed in a sane way.  Each object has its own get_children method to share
everything it carries, while the Base.get_descendants allows all relevant
objects to be found easily.  The various FEproblem.get_* methods will need to
be rewritten to take advantage of this, but this commit is already big enough.

It was also decided to make AxisOrientation objects inherit from Base as well,
thereby being included in the whole descendants structure.  But that means that
an axis can no longer be an arbitrary callable; it must inherit from Base to
behave correctly.  Comments were changed to reflect this, __call__ was changed
to get_at_element, axis_func was changed to axis, and a bunch of tests had to
follow suit.