Replaced FEproblem.get_* methods with a get_descendants_sorted.
This is a little more efficient, as the set of descendants is sorted by type
once.  It effectively trades processing time with storage, but hopefully a few
extra references to already-existing objects won't take any notable amount of
RAM.

The write_feb method was modified accordingly.  The only notable change was
having to modify the materials output code so that TransIsoElastic base
materials are not given their own entries.