Created static "free" and "fixed" constraints.
"Fixed" had previously been a subclass of "Displacement", while a free
constraint was represented by None.  None is an insufficiently explicit way of
representing free, so that was changed to a Force constraint with zero
strength.  And for both free and fixed, there's no reason to have multiple
instances of each.