Finalized approach for contact.
There are currently three contact classes: SlidingContact, TiedContact, and
RigidInterface.  They were created with the idea that the API should focus on
representing the situation to be solved, while details on how to solve it are
passed in as a dictionary of options.

In the case of SlidingContact, the situation is elaborated upon with three
extra parameters outside of the options dict.  There is float
friction_coefficient (if zero, FEBio will use facet-to-facet sliding; else use
sliding_with_gaps), boolean biphasic (if true, FEBio will use sliding2), and
solute (if true, FEBio will use sliding3).  Perhaps these would be better as
separate classes, but for now I consider them extra details on the basic idea
of a sliding contact interface.